backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pyyaml
httpx
python-multipart
pytest
pytest-asyncio
elasticsearch[async]
fastapi-mcp
PyJWT